Unify DAG schedule args and change default to None
The arguments 'schedule_interval' and 'timetable' are removed from both
the DAG class and the `@dag` decorator.

The default value of the 'schedule' argument (on both entities) is
changed to None (i.e. a DAG will not have a schedule by default).

The 'timetable' attribute still exists on DAG, and is now the only value
that reflects the DAG's schedule. The 'schedule_interval' attribute is
removed from DAG.

The 'schedule_interval' on DagModel used to store a string
representation of DAG's attribute of the same name, is now replaced by
timetable_summary, which should (mostly?) work the same as before. We
can fix minor UI differences as we go. Some use cases that rely on that
field are also changed to use other fields instead (dataset_expression,
for example, can be used to check whether a DAG is dataset-triggered).

The API field 'schedule_interval' has also been removed since that field
no longer exists in the database. This has some side effects. The field
previously contains some type information for delta values, but now
becomes a simple string. It is unclear if anyone really needs the
information, but we can always bring it back if needed.