Cleanup installed packages when running provider compatibility tests
When running tests for providers with old Airflow versions, we have
to make sure to uninstall all packages before installing old Airflow
versions - because some of the packages can be installed by new
Airflow version and they are missing in the old Airflow version.

This has already happened in #41402 with methodtools.

This PR adds `--clean-airflow-installation` flag to relevant breeze
commands that install other airflow version. This is quite a bit
slower as it requires to uninstall and reinstall packages so we
do not set it by default.